Replace long double variables by double.
git-svn-id: https://svn.dealii.org/trunk@2755 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/lac/include/lac/solver_cg.h
@@ -195,7 +195,7 @@ SolverCG<Matrix,Vector>::solve (const Matrix &A,
 				   // Implementation taken from the DEAL
 				   // library
   int  it=0;
-  long double res,gh,alpha,beta;
+  double res,gh,alpha,beta;
  
   res = A.residual(g,x,b);
   conv = control().check(0,res);
